Fix Issue in Note cubit and test the add note to hive

+++ lib/models/note_model.dart
@@ -9,7 +9,7 @@ class NoteModel extends HiveObject {
   @HiveField(2)
   final String date;
   @HiveField(3)
-  final int color;
+  final double color;
 
   NoteModel({
     required this.title,
